Rename the heading colors screen file, remove font size for elements.heading.
